packages/auth/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[project]
name = "zion-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "gwainor", email = "me@gokhan.tr" }
]
requires-python = ">=3.12"
dependencies = [
    "annotated-doc>=0.0.4",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.5.0",
    "zion-logger",
    "zion-utils",
]


[project.optional-dependencies]
sqlalchemy = [
    "sqlalchemy>=2.0.46",
    "python-ulid>=3.1.0",
]


[project.urls]
Homepage = "https://github.com/gwainor/zion-python/packages/auth"
Source = "https://github.com/gwainor/zion-python/packages/auth"


[tool.hatch.build.targets.sdist]
include = [
  "/zion_auth",
  "/README.md",
]


[tool.uv.sources]
zion-config = { workspace = true }
zion-utils = { workspace = true }
zion-logger = { workspace = true }
zion-db = { workspace = true }


